refactor(client): NTGL-384 Move logic of timetable tabs to separate file
